Update to the last rust-nigthly.
Version of rustc: 0.12.0-pre-nightly (2e92c67dc 2014-08-28 23:56:20 +0000).

+++ src/window/window.rs
@@ -536,7 +536,7 @@ impl Window {
             Some(ms) => {
                 let elapsed = (time::precise_time_ns() - self.curr_time) / 1000000;
                 if elapsed < ms {
-                    self.timer.sleep(Duration::milliseconds((ms - elapsed) as i32));
+                    self.timer.sleep(Duration::milliseconds((ms - elapsed) as i64));
                 }
             }
         }
